+++ extensions/easydict/src/releaseVersion/changelog.ts
@@ -2,54 +2,42 @@
  * @author: tisfeng
  * @createTime: 2022-07-03 22:10
  * @lastEditor: tisfeng
- * @lastEditTime: 2022-07-04 18:01
+ * @lastEditTime: 2022-07-21 16:03
  * @fileName: changelog.ts
  *
  * Copyright (c) 2022 by tisfeng, All Rights Reserved.
  */
 
 export const changelog = `
-## [1.2.0] - 2022-07-04
+## [v1.3.1] - 2022-07-21
 
-### ✨ 新特性
+### 🐞 Fixes
 
-- 支持苹果🍎系统翻译，需搭配快捷指令食用。[如何在 Easydict 中使用 macOS 苹果系统翻译？](https://github.com/tisfeng/Raycast-Easydict/wiki/%E5%A6%82%E4%BD%95%E5%9C%A8-Easydict-%E4%B8%AD%E4%BD%BF%E7%94%A8-macOS-%F0%9F%8D%8E-%E7%B3%BB%E7%BB%9F%E7%BF%BB%E8%AF%91%EF%BC%9F)
-- 支持苹果🍎系统语言识别。
-- 添加查看最近版本更新日志功能。
-
-![easydict-3](https://i.imgur.com/ycKif2c.jpg)
-
-### 💎 优化
-
-- 改用腾讯，百度，有道等翻译服务自家的 logo ，增加图标辨识度。
-- 优化取词查询的响应速度。
-- 优化自动识别输入文本语言的速度和准确度。
+- 修复了可能多次显示更新日志提示的问题。
+- 修复了有道翻译请求异常错误未处理问题。
 
 ### ✅ 最近更新
 
-- 支持腾讯文本翻译。
-- 支持腾讯语种识别。
-- 支持查询单词后自动播放发音，需在设置中开启。
-- 支持有道 TTS。
+- 支持 DeepL 翻译。
+- 支持手动排序翻译结果显示。
+- 支持苹果🍎系统翻译
+- 修复了删除输入框文本会导致重复查询选中文本问题。
+- 修复了动作按钮渲染时会闪烁显示 “✨ New Version Released” 问题。
+
+![easydict-5](https://raw.githubusercontent.com/tisfeng/ImageBed/main/uPic/easydict-5-1658309785.png)
 
 ---
 
-### ✨ New
+### 🐞 Fixes
 
-- Supported Apple 🍎 system translation.
-- Supported Apple 🍎 system language detection.
-- Added view recent version changelog feature.
-
-### 💎 Improvements
-
-- Used the logo of Youdao, Baidu, Tencent and other translation services to increase icon recognition.
-- Optimized the response speed of querying selected text.
-- Optimized the response speed and accuracy of automatic language detection.
+- fixed new release prompt may be shown multiple times.
+- handled the exceptional error of request to Youdao translation.
 
 ### ✅ Recent Updates
 
-- Supported Tencent text translation.
-- Supported Tencent language detection.
-- Supported automatically play the voice of word after querying, need to enable in settings.
-- Supported Youdao TTS.
+- Added support for DeepL translation.
+- Supported to sort translation results manually.
+- Added support for Apple 🍎 system translation.
+- Fixed deleting input box text will cause repeated query selected text problem.
+- Fixed the bug of "✨ New Version Released" flashing when rendering action button.
 `;