+++ extensions/spotify-player/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Spotify Player Changelog
 
+## [Feature] - 2022-07-21
+
+- Introduce new command "Just Play"
+- Introduce new command "Now Playing" with menu-bar 🔥
+- Introduce new command "Play Similar"
+
 ## [Feature] - 2022-05-24
 
 - Added OAuth Support